clean up pass

+++ examples/resource-runtime-teams.cpp
@@ -159,8 +159,8 @@ int main(int argc, char *argv[])
   RAJA::resources::Resource res = RAJA::expt::Get_Host_Resource(host_res, select_cpu_or_gpu);
 #endif
 
-  const int shared_mem = 0;
-
+  const size_t shared_mem = 0;
+  
   //How the kernel executes now depends on how the resource is constructed (host or device)
   RAJA::expt::launch<launch_policy>
     (res, RAJA::expt::Grid(RAJA::expt::Teams(GRID_SZ),
